+++ changelog.md
@@ -1,8 +1,10 @@
 # Changelog
 
-## 6 IX 2025
-- [ ] Move the SSE architecture [Chores](#chores)
+## 6 IX 2025 - 7 IX 2025
+- [x] Move the SSE architecture [Chores](#chores)
   - [x] Write sensible backend 
 - [ ] Upload to Vercel
 - [ ] Add icon for the app [Misc](#misc)
-- [ ] Implement error handling for the every engine API [Chores](#chores)
+- [x] Implement error handling for the every engine API [Chores](#chores)
+  - [x] Vs-Ai page
+  - [x] Analysis page